put up_then_down with reorder
