+++ src/openfermion/utils/__init__.py
@@ -21,7 +21,8 @@ from ._grid import Grid
 
 from ._operator_utils import (count_qubits, eigenspectrum, fourier_transform,
                               get_file_path, inverse_fourier_transform,
-                              is_identity, load_operator, save_operator,reorder)
+                              is_identity, load_operator, save_operator,
+                              reorder,up_then_down)
 
 from ._slater_determinants import (gaussian_state_preparation_circuit,
                                    slater_determinant_preparation_circuit)
