+++ src/openfermion/utils/_operator_utils.py
@@ -293,7 +293,6 @@ def save_operator(operator, file_name=None, data_directory=None,
 
 
 def reorder(Operator, order_function,num_modes=None, reverse=False):
-
     """ changes the fermionic order of the Hamiltonian based on the provided
     order_function per mode index
     Args:
@@ -307,6 +306,8 @@ def reorder(Operator, order_function,num_modes=None, reverse=False):
             calculated based on the Operator.
         reverse (bool): default False. if set to True, the mode mapping is
             reversed
+
+    Note: Every order function must take in a mode_idx and num_modes.
     """
 
     if num_modes is None:
@@ -323,3 +324,27 @@ def reorder(Operator, order_function,num_modes=None, reverse=False):
         new_term = tuple([(mode_map[op[0]], op[1]) for op in term])
         rotated_hamiltonian += Operator.__class__(new_term, value)
     return rotated_hamiltonian
+
+
+def up_then_down(mode_idx, num_modes):
+    """ up then down reordering, given the operator has the default even-odd
+     ordering. Otherwise this function will reorder indices where all even
+     indices now come before odd indices.
+
+     Example:
+         0,1,2,3,4,5 -> 0,2,4,1,3,5
+
+    The function takes in the index of the mode that will be relabeled and
+    the total number modes.
+
+    Args:
+        mode_idx (int): the mode index that is being reordered
+        num_modes (int): the total number of modes of the operator.
+
+    Returns (int): reordered index of the mode.
+    """
+    halfway = num_modes / 2
+    if mode_idx % 2 == 0:
+        return mode_idx // 2
+    else:
+        return mode_idx // 2 + halfway
